requirements.txt
streamlit==1.42.0
plotly==5.24.0
ccxt==4.4.92
pandas==2.2.3
numpy==2.2.1
python-dotenv==1.0.1
APScheduler==3.10.4
sqlalchemy==2.0.38
aiosqlite==0.20.0
pydantic==2.10.6
requests==2.32.3
httpx==0.27.2
python-telegram-bot==21.9
